refactor: package DSL agent core modules
