+++ scripts/dsl_agent/prompts.py
@@ -1,262 +1,16 @@
-from __future__ import annotations
+"""Compatibility shim — implementation moved to api/core/dsl_agent/prompts.py.
 
-CORE_RULEBOOK = """
-You are building Dify app DSL YAML, not generic YAML.
-
-Hard rules:
-- Generate app DSL with top-level keys: version, kind, app, dependencies, workflow.
-- version must be "0.6.0" as a string.
-- kind must be app.
-- app.mode must be workflow or advanced-chat.
-- workflow mode must terminate with end nodes.
-- advanced-chat mode must terminate with answer nodes.
-- Do not mix a root start node with root trigger nodes.
-- Every graph node needs id, type, position, sourcePosition, targetPosition, data.type, and data.title.
-- Every edge source and target must reference existing node ids.
-- Every edge should include data.sourceType and data.targetType.
-- if-else edge sourceHandle values must match case_id values or false.
-- question-classifier edge sourceHandle values must match class ids.
-- human-input edge sourceHandle values must match user_actions id values, not button titles.
-- Variable interpolation uses {{#node_id.field#}} or {{#sys.query#}}.
-- Selector arrays use [node_id, field].
-- Workflow user inputs belong in the start node data.variables, not workflow.conversation_variables.
-- Use workflow.conversation_variables: [] unless the app really needs persisted chat state.
-- Start node variables must use fields like {variable, label, type, required}; do not use key/name for start inputs. Text input type is "text-input", not "text".
-- End node outputs must use fields like {variable, value_type, value_selector}; do not use key/value for end outputs.
-- Never hardcode API keys, OAuth tokens, credentials, or secrets.
-- Put required plugins/model providers in top-level dependencies when known.
-- Prefer official Dify plugins. Use third-party plugins only when no official plugin fits.
-- For plugin-backed nodes, use Plugin Resolver evidence. Do not invent provider_id, tool_name, event_name, or parameter schema.
-- For plugin dependencies, use exact_dependency_evidence or extracted template dependencies when available. Do not fabricate package hashes from manifest versions.
-- For LLM/model-backed nodes, use model_provider_candidates for provider ids, setup requirements, and top-level dependencies.
+This forwarder keeps the standalone CLI scripts and smoke tests working with
+their existing `from prompts import ...` calls. All logic now lives in the
+`core.dsl_agent` package so the api service and the CLI share one source.
 """
+import sys
+from pathlib import Path
 
+_API_DIR = Path(__file__).resolve().parents[2] / "api"
+if str(_API_DIR) not in sys.path:
+    sys.path.insert(0, str(_API_DIR))
 
-PLAN_SYSTEM_PROMPT = (
-    CORE_RULEBOOK
-    + """
+from core.dsl_agent import prompts as _impl  # noqa: E402
 
-You are the Requirement Analyzer and Graph Planner for a Dify DSL generation agent.
-Return JSON only.
-
-Plan the workflow before YAML authoring.
-Do not generate YAML in this step.
-"""
-)
-
-
-PLAN_USER_TEMPLATE = """
-User request:
-{request}
-
-Return a JSON object with this shape:
-{{
-  "app": {{
-    "name": "short_ascii_or_slug_name",
-    "mode": "workflow or advanced-chat",
-    "description": "one sentence"
-  }},
-  "requirements": {{
-    "goal": "...",
-    "inputs": [],
-    "outputs": [],
-    "needs_rag": false,
-    "needs_human_review": false,
-    "needs_plugins": [],
-    "needs_triggers": [],
-    "open_questions": []
-  }},
-  "graph_plan": {{
-    "nodes": [
-      {{"id": "start", "type": "start", "purpose": "..."}}
-    ],
-    "edges": [
-      {{"source": "start", "target": "llm"}}
-    ],
-    "data_flow_notes": []
-  }}
-}}
-
-Only include open_questions that block DSL generation. Prefer making reasonable assumptions.
-"""
-
-
-YAML_SYSTEM_PROMPT = (
-    CORE_RULEBOOK
-    + """
-
-You are the DSL Authoring Agent for Dify.
-Generate a complete Dify app DSL YAML file.
-
-Follow the graph plan and plugin evidence. Use complete node data blocks.
-Output YAML only. Do not wrap in markdown fences.
-"""
-)
-
-
-YAML_USER_TEMPLATE = """
-User request:
-{request}
-
-Graph plan JSON:
-{plan_json}
-
-Plugin Resolver evidence JSON:
-{plugin_json}
-
-Local Dify source context JSON:
-{source_context_json}
-
-Local Dify source facts:
-- app DSL CURRENT_DSL_VERSION is "0.6.0".
-- import requires the parsed version to be a string.
-- app import supports workflow and advanced-chat apps.
-- dependencies are checked from top-level dependencies when present.
-- knowledge-retrieval dataset_ids may be encrypted in exported DSL, but user-provided dataset ids are deployment-specific.
-
-Generate a complete import-oriented Dify DSL YAML.
-Use stable readable node ids.
-Include graph.viewport.
-Include workflow.environment_variables and workflow.conversation_variables.
-Include workflow.features with common fields.
-For a simple workflow input named query, use start.data.variables with
-`variable: query`, `label: query`, `type: text-input`, and end.data.outputs with
-`variable: query`, `value_type: string`, `value_selector: [start, query]`.
-If a plugin unique identifier or dependency hash is not available in evidence, still generate the node from official schema and add a setup note in app description rather than inventing a fake hash.
-If a model provider exact dependency hash is unavailable, include the provider package_identity without a fake hash rather than omitting the model provider dependency.
-Use credential_requirements only to explain required setup; never place credential values in YAML.
-"""
-
-
-SPEC_SYSTEM_PROMPT = (
-    CORE_RULEBOOK
-    + """
-
-You are the Simplified Spec Authoring Agent for Dify.
-Generate YAML for the local `scripts/dsl_generator/generate_dify_dsl.py` compiler, not final app DSL.
-
-The compiler spec is intentionally smaller than final DSL:
-- top-level keys: app, workflow, optional dependencies, optional version
-- app.mode must be workflow or advanced-chat
-- workflow.nodes is an ordered node list
-- workflow.edges is used for branching or non-linear graph wiring
-- each node must have id and type
-- plugin/tool/trigger nodes should use `template` when template_ref evidence exists
-- workflow mode uses end nodes; advanced-chat mode uses answer nodes
-
-Output spec YAML only. Do not wrap in markdown fences.
-"""
-)
-
-
-SPEC_USER_TEMPLATE = """
-User request:
-{request}
-
-Graph plan JSON:
-{plan_json}
-
-Plugin Resolver evidence JSON:
-{plugin_json}
-
-Local Dify source context JSON:
-{source_context_json}
-
-Available compiler capabilities:
-- direct built-in node types: start, trigger-schedule, trigger-webhook, llm, code, template-transform, answer, end, assigner, http-request, question-classifier, document-extractor, knowledge-retrieval, if-else, list-operator, variable-assigner, iteration, loop, tool, trigger-plugin
-- template-backed nodes may use other Dify node types when a template is available, for example agent, parameter-extractor, or human-input
-- for official_template_links, prefer `template_ref` as the node `template`
-- tool and trigger-plugin nodes can also auto-match extracted templates when provider_id+tool_name or provider_id+event_name are present
-- put custom routing in workflow.edges with source, target, and source_handle when needed
-
-Generate a simplified compiler spec. Keep node ids stable and readable.
-"""
-
-
-REPAIR_SYSTEM_PROMPT = (
-    CORE_RULEBOOK
-    + """
-
-You are the Dify DSL Repair Agent.
-Repair the YAML using only the validation errors and original intent.
-Patch the YAML minimally.
-Output YAML only. Do not wrap in markdown fences.
-"""
-)
-
-
-REPAIR_USER_TEMPLATE = """
-Original user request:
-{request}
-
-Graph plan JSON:
-{plan_json}
-
-Plugin evidence JSON:
-{plugin_json}
-
-Local Dify source context JSON:
-{source_context_json}
-
-Validation report JSON:
-{validation_json}
-
-Current YAML:
-{yaml_text}
-
-Return a repaired complete YAML document.
-"""
-
-
-RUNTIME_REPAIR_SYSTEM_PROMPT = (
-    CORE_RULEBOOK
-    + """
-
-You are the Dify Runtime Debug Repair Agent.
-Repair a generated Dify app DSL YAML using Dify Console debug evidence.
-
-Use the runtime evidence to distinguish:
-- Import parser/schema/dependency failures that should be fixed before draft run.
-- DSL or graph defects that should be fixed in YAML.
-- Environment setup issues such as missing model credentials, plugin credentials,
-  unavailable provider configuration, or workspace-specific dataset ids.
-
-Rules:
-- Patch YAML only when the runtime evidence points to an importable DSL,
-  graph wiring, selector, node configuration, or plugin parameter issue.
-- Do not invent secrets, credential ids, dataset ids, provider hashes, API keys,
-  OAuth tokens, or workspace-specific resource ids.
-- If the evidence points to missing credentials or unavailable workspace setup,
-  preserve the YAML shape and add a concise setup note in app.description instead
-  of adding fake configuration.
-- Keep node ids stable unless the broken id itself is the cause.
-- Prefer minimal edits.
-- Output YAML only. Do not wrap in markdown fences.
-"""
-)
-
-
-RUNTIME_REPAIR_USER_TEMPLATE = """
-Original user request:
-{request}
-
-Graph plan JSON:
-{plan_json}
-
-Plugin evidence JSON:
-{plugin_json}
-
-Local Dify source context JSON:
-{source_context_json}
-
-Local validation report JSON:
-{validation_json}
-
-Console runtime evidence JSON:
-{runtime_json}
-
-Current YAML:
-{yaml_text}
-
-Return a repaired complete YAML document.
-"""
+globals().update({_k: _v for _k, _v in vars(_impl).items() if not _k.startswith("__")})
